add notifications and confirmation modal with sweetalert2

+++ src/modules/daybook/views/EntryView.vue
@@ -36,6 +36,7 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'
+import Swal from 'sweetalert2'
 
 import getDayMonthYear from '../helpers/getDayMonthYear'
 
@@ -90,20 +91,41 @@ export default {
       this.entry = entry
     },
     async saveEntry() {
+      new Swal({
+        title: 'Espere por favor',
+        allowOutsideClick: false
+      })
+      Swal.showLoading()
+
       if (this.entry.id) {
         await this.updateEntry(this.entry)
       } else {
         const id = await this.createEntry(this.entry)
-
-        console.log({ id })
-
         this.$router.push({ name: 'entry', params: { id } })
       }
+
+      Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
     },
     async onDeleteEntry() {
-      await this.deleteEntry(this.entry.id)
+      const { isConfirmed } = await Swal.fire({
+        title: '¿Está seguro?',
+        text: 'Una vez borrado, no se puede recuperar',
+        showDenyButton: true,
+        confirmButtonText: 'Sí, estoy seguro'
+      })
 
-      this.$router.push({ name: 'no-entry' })
+      if (isConfirmed) {
+        new Swal({
+          title: 'Espere por favor',
+          allowOutsideClick: false
+        })
+        Swal.showLoading()
+
+        await this.deleteEntry(this.entry.id)
+        this.$router.push({ name: 'no-entry' })
+
+        Swal.fire('Eliminado', '', 'success')
+      }
     }
   },
